Created README.md + Please Read Note
Changes/Updates: Originally there were 3 files in the Java PSS with
redundant code throughout. CSV reading and writing functions have been
moved to their own files. Beyond that there is generally less bloat,
redundant code, and I found that my calculating of the salt value was
overly complicated in the original commit. The only other major update
is that I didn't notice the PSSDriver for the JFree and Apache
associated files still had a method in it that was not the main method
so that was also moved to its own file.
The test folder is an old folder I used to test my IDE to git connection
Finally, the edit made to this program were made over the last few days if for nothing else then it was because I noticed the mistakes and in an attempt to grow as a coder I aimed to fix them. If Professor Hoy is reading this and you haven't graded my project yet and do not accept these edits please go back to the older commit that was turned in on time.